rename: Components 구조 설계 후 변경
