+++ front/src/components/Content/CardWritable/CardWritable.js
