+++ front/src/components/Content/Content.js
@@ -1,11 +1,11 @@
-import peact from "../core/peact";
+import peact from "../../core/peact";
 import Columns from "./Column/Columns";
 
-const Main = ({ columns, todos }) => {
+const Content = ({ columns, todos }) => {
   return peact.createElement({
     tag: "div",
     child: [Columns({ columns, todos })],
   });
 };
 
-export default Main;
+export default Content;
