--- front/src/components/main.module.css
@@ -1,1 +1,0 @@
-
